requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
alembic
pydantic
httpx
python-jose
passlib
bcrypt==4.0.1
PyJWT
python-dotenv
python-multipart
cryptography
